python_proj/requirements.txt
numpy==1.19.2
Flask_RESTful==0.3.8
statsmodels==0.12.0
Flask==1.1.2
atomos==0.3.1
netifaces==0.10.9
pandas==1.1.2
etcd3==0.12.0
cassandra_driver==3.24.0
